requirements.txt
fastapi            
uvicorn           
oracledb           
pydantic             
pydantic-settings      
streamlit              
python-dotenv        
pandas                 
sqlalchemy            
black                 
isort    
Pillow        
moviepy
opencv-python
ffmpeg

